Updates to rstl; rename some symbols; update CFLAGS

+++ include/types.h
@@ -7,12 +7,12 @@ extern "C" {
 
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned long u32;
+typedef unsigned int u32;
 typedef unsigned long long u64;
 
 typedef char s8;
 typedef short s16;
-typedef long s32;
+typedef int s32;
 typedef long long s64;
 
 typedef volatile u8 vu8;
